Abording / Pausing Network Project
Too Complicated - every object needs to be handled on it's own via network.
Would have been cool if there was a general way to synchronize worlds - or if Networking was in the projects mind from the beginning. It's too advanced now and hell to put everything into network.
Also: Can't send rpc commands over network when a Object is used as parameter - only common types are allowed. So, working mostly with ID's here (which makes up for ALOT of changes in the code).
Would be easier if there also was a strict border between model and view - but some parts of the model rely on various parts of the view to complete.
It's possible, just takes alot of work.
Implemented, but aborded:
-Chat Feature (works fully)
-some movement stuff
-item spawn
-star collecting

Focus will now be on local - more levels.